+++ pkg/runtime/br.go
@@ -23,6 +23,7 @@ import (
 	brv1alpha1 "github.com/pingcap/tidb-operator/api/v2/br/v1alpha1"
 	metav1alpha1 "github.com/pingcap/tidb-operator/api/v2/meta/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/client"
+	"github.com/pingcap/tidb-operator/tests/e2e/utils/k8s"
 )
 
 // Job is the interface for a job. for example, backup, restore, etc.
@@ -98,23 +99,10 @@ func (b *Backup) NeedRemoveFinalizer() bool {
 }
 
 func isBackupDeletionCandidate(backup *v1alpha1.Backup) bool {
-	return backup.DeletionTimestamp != nil && ContainsString(backup.Finalizers, metav1alpha1.Finalizer, nil)
+	return backup.DeletionTimestamp != nil && k8s.ContainsString(backup.Finalizers, metav1alpha1.Finalizer, nil)
 }
 
 // TODO(ideascf): move this to pkg/utils/k8s
-// ContainsString checks if a given slice of strings contains the provided string.
-// If a modifier func is provided, it is called with the slice item before the comparation.
-func ContainsString(slice []string, s string, modifier func(s string) string) bool {
-	for _, item := range slice {
-		if item == s {
-			return true
-		}
-		if modifier != nil && modifier(item) == s {
-			return true
-		}
-	}
-	return false
-}
 
 func (b *Backup) Completed() bool {
 	return brv1alpha1.IsBackupComplete(ToBackup(b))
@@ -169,14 +157,12 @@ func (b *Backup) SetConditions(conds []metav1.Condition) {
 }
 
 func (b *Backup) ObservedGeneration() int64 {
-	// return b.Status.ObservedGeneration
-	// TODO(ideascf): do we need this?
+	// dummy function
 	return 0
 }
 
 func (b *Backup) SetObservedGeneration(g int64) {
-	// TODO(ideascf): do we need this?
-	// b.Status.ObservedGeneration = g
+	// dummy function
 }
 
 func (b *Backup) Object() client.Object {
@@ -233,14 +219,12 @@ func (r *Restore) SetConditions(conds []metav1.Condition) {
 }
 
 func (r *Restore) ObservedGeneration() int64 {
-	// return r.Status.ObservedGeneration
-	// TODO(ideascf): do we need this?
+	// dummy function
 	return 0
 }
 
 func (r *Restore) SetObservedGeneration(g int64) {
-	// TODO(ideascf): do we need this?
-	// r.Status.ObservedGeneration = g
+	// dummy function
 }
 
 func (r *Restore) Object() client.Object {
