+++ tests/e2e/utils/k8s/k8s.go
@@ -123,3 +123,17 @@ func getTimeFromEvent(e watch.Event) time.Time {
 	}
 	return time.Time{}
 }
+
+// ContainsString checks if a given slice of strings contains the provided string.
+// If a modifier func is provided, it is called with the slice item before the comparation.
+func ContainsString(slice []string, s string, modifier func(s string) string) bool {
+	for _, item := range slice {
+		if item == s {
+			return true
+		}
+		if modifier != nil && modifier(item) == s {
+			return true
+		}
+	}
+	return false
+}
